requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.9.0
altair<5
sentence-transformers==2.2.2
boto3==1.35.66
openpyxl==3.1.5
flask==3.1.0